Fix lib import and componentDidMount bugs
