+++ src/views/components/SVG.jsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import Utils from '../../lib/danehansen/Utils';
+import Utils from '../../lib/danehansen/utils/Utils';
 
 const _NS = 'http://www.w3.org/2000/svg';
 
 class SVG extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {};
+    this.state = {
+      touch: false,
+    };
   }
 
   render() {
@@ -16,8 +18,8 @@ class SVG extends React.Component {
       var move = this.props.move;
       var out = this.props.out;
       return (
-        <svg className={this.props.className} version='1.1' xmlns={_NS} x='0px' y='0px' width={width+'px'} height={height+'px'} viewBox={'0 0 '+width+' '+height} onMouseMove={this._touch ? null : move}
-          onTouchMove={this._touch ? move : null} onMouseLeave={this._touch ? null : out} onTouchEnd={this._touch ? out : null}>
+        <svg className={this.props.className} version='1.1' xmlns={_NS} x='0px' y='0px' width={width+'px'} height={height+'px'} viewBox={'0 0 '+width+' '+height} onMouseMove={this.state.touch ? null : move}
+          onTouchMove={this.state.touch ? move : null} onMouseLeave={this.state.touch ? null : out} onTouchEnd={this.state.touch ? out : null}>
           {this.props.children}
         </svg>
       );
@@ -43,7 +45,7 @@ class SVG extends React.Component {
   }
 
   componentDidMount() {
-    this._touch = Utils.touch();
+    this.setState({touch: Utils.touch()});
   }
 }
