+++ src/views/components/icons/RIcon.jsx
@@ -19,7 +19,7 @@ class RIcon extends React.Component {
     return (
       <SVG className='SVG-icon RIcon' width={SVG.ICON_SIZE} height={SVG.ICON_SIZE} fallbackIcon='icon-r-circled'>
         <circle className='SVG-fill' cx={SVG.ICON_SIZE / 2} cy={SVG.ICON_SIZE / 2} r={SVG.ICON_SIZE / 2}/>
-        <path ref='r' className={'SVG-fill-bg' + (this.state.outline?' outline':'')} d="M12.198584,5.704752c-1.356266,0-2.418676,0.421916-3.200488,1.089559 C8.646981,6.258162,8.041927,5.903345,7.35319,5.903345H7.333333v4.765951v3.177311v0.164551 c0,0.528857,0.399096,0.990633,0.926847,1.024789c0.578288,0.037427,1.058969-0.420573,1.058969-0.990764v-0.198576v-3.673763 c0-1.370922,1.111352-2.482275,2.482267-2.482275c0.822558,0.000008,1.489364-0.666805,1.489364-1.489356V5.704761 L12.198584,5.704752z"/>
+        <path ref='r' className={'SVG-fill-bg' + (this.state.outline?' outline':'')} d='M7.333333,5.984375v8.026783c0,0.528857,0.399096,0.990633,0.926847,1.024789 c0.578288,0.037427,1.058969-0.420573,1.058969-0.990764v-3.872339c0-1.370922,1.111352-2.482275,2.482267-2.482275 c0.822558,0.000008,1.489364-0.666805,1.489364-1.489356V5.704761l-1.092196-0.000009 c-1.356266,0-2.418676,0.421916-3.200488,1.089559C8.646981,6.258162,8.041927,5.903345,7.35319,5.903345H7.333333V5.984375z'/>
       </SVG>
     );
   }
